Combine Main.java into Launcher.java
