+++ src/main/java/duke/main/DialogBox.java
@@ -7,13 +7,12 @@ import javafx.scene.Node;
 import javafx.scene.control.Label;
 import javafx.scene.image.ImageView;
 import javafx.scene.layout.HBox;
+import javafx.scene.shape.Circle;
 
 public class DialogBox extends HBox {
 
     private Label text;
     private ImageView displayPicture;
-
-
 
     public DialogBox(Label l, ImageView iv) {
         text = l;
@@ -23,7 +22,31 @@ public class DialogBox extends HBox {
         displayPicture.setFitWidth(100.0);
         displayPicture.setFitHeight(100.0);
 
+        Circle clip = new Circle(50, 50, 50); // Centered at (50, 50) with a radius of 50
+        displayPicture.setClip(clip); // Set the clip for the ImageView
+
+//        HBox content = new HBox(0); // 10 pixels spacing between ImageView and Label
+//        content.getChildren().addAll(displayPicture, text);
+//        content.setStyle("-fx-background-color: #726060; -fx-padding: 0px;");
+
         this.setAlignment(Pos.TOP_RIGHT);
         this.getChildren().addAll(text, displayPicture);
+    }
+
+    private void flip() {
+        this.setAlignment(Pos.TOP_LEFT);
+        ObservableList<Node> tmp = FXCollections.observableArrayList(this.getChildren());
+        FXCollections.reverse(tmp);
+        this.getChildren().setAll(tmp);
+    }
+
+    public static DialogBox getUserDialog(Label l, ImageView iv) {
+        return new DialogBox(l, iv);
+    }
+
+    public static DialogBox getDukeDialog(Label l, ImageView iv) {
+        var db = new DialogBox(l, iv);
+        db.flip();
+        return db;
     }
 }
